fix: synchronize patching binding/createWriteStream/cwd/chdir
All the patches now check same realBinding._mockedBinding. This should
fix the edge cases around the new feature bypass().

follows up #306
