+++ lib/index.js
@@ -87,7 +87,9 @@ function overrideCreateWriteStream() {
   fs.createWriteStream = function(path, options) {
     const output = realCreateWriteStream(path, options);
     // disable _writev, this will over shadow WriteStream.prototype._writev
-    output._writev = undefined;
+    if (realBinding._mockedBinding) {
+      output._writev = undefined;
+    }
     return output;
   };
 }
@@ -126,13 +128,20 @@ exports = module.exports = function mock(config, options) {
   let currentPath = process.cwd();
   overrideProcess(
     function cwd() {
-      return currentPath;
+      if (realBinding._mockedBinding) {
+        return currentPath;
+      }
+      return realProcessProps.cwd();
     },
     function chdir(directory) {
-      if (!binding.stat(toNamespacedPath(directory)).isDirectory()) {
-        throw new FSError('ENOTDIR');
+      if (realBinding._mockedBinding) {
+        if (!binding.stat(toNamespacedPath(directory)).isDirectory()) {
+          throw new FSError('ENOTDIR');
+        }
+        currentPath = path.resolve(currentPath, directory);
+      } else {
+        return realProcessProps.chdir(directory);
       }
-      currentPath = path.resolve(currentPath, directory);
     }
   );
